requirements.txt
django_haystack==3.2.1
docxtpl==0.16.7
jieba==0.42.1
Pillow==9.5.0
pyquery==2.0.0
python_docx==1.0.1
qrcode==7.4.2
six==1.16.0
South==1.0.2
Whoosh==2.7.4
xadmin==0.6.1
